requirements.txt
litellm>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest-asyncio>=0.23.0
